Fix compiler warnings in Psyco.
git-svn-id: http://unladen-swallow.googlecode.com/svn/trunk@108 05521daa-c0b4-11dd-bb00-bd6ab96fe29a

+++ Modules/_psyco/i386/ipyencoding.h
@@ -138,7 +138,7 @@ PSY_INLINE void dictitem_update_nochange(void* originalmacrocode,
   NEED_CC();                                            \
   code[0] = 0xFF;  /* INC [address] */                  \
   code[1] = 0x05;                                       \
-  *(int**)(code+2) = &(pyobj)->ob_refcnt;               \
+  *(Py_ssize_t**)(code+2) = &(pyobj)->ob_refcnt;        \
   code += 6;                                            \
  } while (0)
 
@@ -147,7 +147,7 @@ PSY_INLINE void dictitem_update_nochange(void* originalmacrocode,
   NEED_CC();                                            \
   code[0] = 0xFF;  /* DEC [address] */                  \
   code[1] = (1<<3) | 0x05;                              \
-  *(int**)(code+2) = &(pyobj)->ob_refcnt;               \
+  *(Py_ssize_t**)(code+2) = &(pyobj)->ob_refcnt;        \
   code += 6;                                            \
  } while (0)
